Params: Add option for getssl_dir
This commit allows for non-standard getssl config/data directory locations
to be used for certificate updates by parsing argv[3] as a string pointing
to getssl's directory.

Trailing slashes is taken care of by calling rstrip('/') on argv[3]

Note the fallback, the default value for getssl_dir remains to be
'/root/.getssl', so we keep 100% backwards compatibility.